tests/wpt/web-platform-tests/css/css-transforms/animation/perspective-interpolation.html

<!DOCTYPE html>
<meta charset="utf-8">
<title> perspective interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-perspective">
<meta name="assert" content="perspective supports animation">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
.parent {
  perspective: 30px;
}
.target {
  perspective: 10px;
}
.transformed {
  width: 50px;
  height: 50px;
  background: black;
  transform: rotateY(45deg);
}
.expected .transformed {
  background: green;
}
.expected {
  position: relative;
  left: -50px;
  opacity: 0.75;
}
</style>
<body>
<template id="target-template">
<div><div class="transformed"></div></div>
</template>
<script>
test_interpolation({
  property: 'perspective',
  from: neutralKeyframe,
  to: '20px',
}, [
  {at: -20, expect: 'none'},
  {at: -1, expect: 'none'},
  {at: -0.3, expect: '7px'},
  {at: 0, expect: '10px'},
  {at: 0.3, expect: '13px'},
  {at: 0.6, expect: '16px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '25px'},
]);

test_no_interpolation({
  property: 'perspective',
  from: 'initial',
  to: '20px',
});

test_interpolation({
  property: 'perspective',
  from: 'inherit',
  to: '20px',
}, [
  {at: -20, expect: '230px'},
  {at: -1, expect: '40px'},
  {at: -0.3, expect: '33px'},
  {at: 0, expect: '30px'},
  {at: 0.3, expect: '27px'},
  {at: 0.6, expect: '24px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '15px'},
]);

test_no_interpolation({
  property: 'perspective',
  from: 'unset',
  to: '20px',
});

test_interpolation({
  property: 'perspective',
  from: '50px',
  to: '100px',
}, [
  {at: -20, expect: 'none'}, // perspective does not accept 0 or negative values
  {at: -1, expect: 'none'}, // perspective does not accept 0 or negative values
  {at: -0.3, expect: '35px'},
  {at: 0, expect: '50px'},
  {at: 0.3, expect: '65px'},
  {at: 0.6, expect: '80px'},
  {at: 1, expect: '100px'},
  {at: 1.5, expect: '125px'},
]);

test_no_interpolation({
  property: 'perspective',
  from: '50px',
  to: 'none',
});
</script>
</body>
